feature store handler of Google Health COVID-19 Open Data features

This DiseasystoreGoogleCovid19 R6 brings support for using the Google Health COVID-19 Open Data repository. See the vignette("diseasystore-google-covid-19") for details on how to configure the feature store.

Value

A new instance of the DiseasystoreGoogleCovid19 R6 class.

Super class

diseasystore::DiseasystoreBase -> DiseasystoreGoogleCovid19

Methods

Public methods

DiseasystoreGoogleCovid19$clone()

Inherited methods

diseasystore::DiseasystoreBase$get_feature()

diseasystore::DiseasystoreBase$initialize()

diseasystore::DiseasystoreBase$key_join_features()

Method clone()

The objects of this class are cloneable with this method.

Usage

DiseasystoreGoogleCovid19$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  ds <- DiseasystoreGoogleCovid19$new(
    source_conn = ".",
    target_conn = DBI::dbConnect(RSQLite::SQLite())
  )

  rm(ds)