diseasystore: Feature Stores for the 'diseasy' Framework

Simple feature stores and tools for creating personalised feature stores. 'diseasystore' powers feature stores which can automatically link and aggregate features to a given stratification level. These feature stores are automatically time-versioned (powered by the 'SCDB' package) and allows you to easily and dynamically compute features as part of your continuous integration.

See also

Author

Maintainer: Rasmus Skytte Randløv rske@ssi.dk (ORCID)

Other contributors:

Marcus Munch Grünewald (ORCID) [contributor]

Kaare Græsbøll kagr@ssi.dk (ORCID) [reviewer]

Kasper Schou Telkamp (ORCID) [reviewer]

Lasse Engbo Christiansen lsec@ssi.dk (ORCID) [reviewer]

Sofia Myrup Otero smot@ssi.dk [reviewer]

Statens Serum Institut, SSI [copyright holder, funder]