Create an object for logging database operations

Create an object for logging database operations

Create an object for logging database operations

Create an object for logging database operations

Create an object for logging database operations

Methods

Public methods

mg_Logger$new()

mg_Logger$generate_filename()

mg_Logger$generate_log_entry()

mg_Logger$log_format()

mg_Logger$log_info()

mg_Logger$log_warn()

mg_Logger$log_error()

mg_Logger$log_to_db()

mg_Logger$clone()

Method new()

Usage

mg_Logger$new(
  db_tablestring = NULL,
  log_db = getOption("mg.log_db"),
  log_conn = NULL,
  log_path = getOption("mg.log_path"),
  ts = NULL,
  tic = Sys.time()
)

Arguments

log_db

The specification to where db logs are stored ("schema.table"). If NULL, no db logs are created.

log_path

Path to a directory of log files.

log_path

The path where logs are stored. If NULL, no file logs are created.

Method generate_filename()

Usage

mg_Logger$generate_filename(db_tablestring = self$db_tablestring, ts = self$ts)

Method generate_log_entry()

Usage

mg_Logger$generate_log_entry()

Method log_format()

Usage

mg_Logger$log_format(..., tic = NULL, log_type = NULL)

Method log_info()

Usage

mg_Logger$log_info(..., tic = Sys.time(), log_type = "INFO")

Method log_warn()

Usage

mg_Logger$log_warn(..., log_type = "WARNING")

Method log_error()

Usage

mg_Logger$log_error(..., log_type = "ERROR")

Method log_to_db()

Usage

mg_Logger$log_to_db(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

mg_Logger$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Methods

Public methods

mg_Logger$new()

mg_Logger$generate_filename()

mg_Logger$generate_log_entry()

mg_Logger$log_format()

mg_Logger$log_info()

mg_Logger$log_warn()

mg_Logger$log_error()

mg_Logger$log_to_db()

mg_Logger$clone()

Method new()

Usage

mg_Logger$new(
  db_tablestring = NULL,
  log_db = getOption("mg.log_db"),
  log_conn = NULL,
  log_path = getOption("mg.log_path"),
  ts = NULL,
  tic = Sys.time()
)

Arguments

log_db

The specification to where db logs are stored ("schema.table"). If NULL, no db logs are created.

log_path

Path to a directory of log files.

log_path

The path where logs are stored. If NULL, no file logs are created.

Method generate_filename()

Usage

mg_Logger$generate_filename(db_tablestring = self$db_tablestring, ts = self$ts)

Method generate_log_entry()

Usage

mg_Logger$generate_log_entry()

Method log_format()

Usage

mg_Logger$log_format(..., tic = NULL, log_type = NULL)

Method log_info()

Usage

mg_Logger$log_info(..., tic = Sys.time(), log_type = "INFO")

Method log_warn()

Usage

mg_Logger$log_warn(..., log_type = "WARNING")

Method log_error()

Usage

mg_Logger$log_error(..., log_type = "ERROR")

Method log_to_db()

Usage

mg_Logger$log_to_db(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

mg_Logger$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.