Opens connection to the database

Connects to the specified dbname of host:port using user and password from given arguments. Certain drivers may use credentials stored in a file, such as ~/.pgpass (PostgreSQL)

Overloads the dplyr _join to accept an na_by argument. By default, joining using SQL does not match on NA / NULL. dbplyr has the option "na_matches = na" to match on NA / NULL but this is very inefficient This function does the matching more efficiently. If a column contains NA / NULL, give the argument to na_by to match during the join If no na_by is given, the function defaults to using dplyr::_join

The function "interlaces" the queries and combines their validity time axes onto a single time axis

If filter = NULL, not filtering is done If filter is different from NULL, the .data is filtered by a mg_inner_join using all columns of the filter: mg_inner_join(.data, filter, by = colnames(filter))

by and na_by can overwrite the mg_inner_join columns used in the filtering

Create an object for logging database operations

Usage

mg_get_connection(
  drv = RPostgres::Postgres(),
  host = NULL,
  port = NULL,
  dbname = NULL,
  user = NULL,
  password = NULL,
  timezone = NULL,
  timezone_out = NULL,
  ...
)

mg_get_tables(conn, pattern = NULL)

mg_get_table(
  conn,
  db_table_id = NULL,
  slice_ts = NA,
  include_slice_info = FALSE
)

mg_table_exists(conn, db_table_id)

mg_id(db_table_id, conn = NULL)

mg_slice_time(.data, slice_ts, from_ts = from_ts, until_ts = until_ts)

mg_inner_join(x, y, by = NULL, na_by = NULL, ...)

mg_left_join(x, y, by = NULL, na_by = NULL, ...)

mg_mg_right_join(x, y, sql_on, renamer, ...)

# S3 method for tbl_dbi
mg_mg_right_join(x, y, sql_on, renamer, ...)

# S3 method for tbl_SQLiteConnection
mg_mg_right_join(x, y, sql_on, renamer, ...)

mg_right_join(x, y, by = NULL, na_by = NULL, ...)

mg_join_na_sql(by, na_by)

mg_select_na_sql(x, y, by, na_by, left = TRUE)

mg_join_warn()

mg_join_warn_experimental()

mg_interlace_sql(tables, by = NULL, colnames = NULL)

mg_digest_to_checksum(.data, col = "checksum", exclude = NULL, warn = TRUE)

mg_digest_to_checksum_native_md5(.data, col)

mg_digest_to_checksum_internal(.data, col)

# S3 method for default
mg_digest_to_checksum_internal(.data, col)

# S3 method for tbl_dbi
unite(data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)

mg_create_table(.data, conn = NULL, db_table_id = NULL, temporary = TRUE, ...)

mg_update_snapshot(
  .data,
  conn,
  db_table,
  timestamp,
  filters = NULL,
  message = NULL,
  tic = Sys.time(),
  log_path = getOption("mg.log_path"),
  log_table_id = getOption("mg.log_table_id"),
  enforce_chronological_order = TRUE
)

mg_filter_keys(.data, filters, by = NULL, na_by = NULL)

mg_db_timestamp(timestamp, conn)

mg_is.historical(.data)

mg_nrow(.data)

mg_create_logs_if_missing(log_table, conn)

mg_age_labels(age_cuts)

mg_assert_dbtable_like(db_table, ..., add = NULL)

mg_assert_timestamp_like(timestamp, ..., add = NULL)

mg_assert_data_like(.data, ..., add = NULL)

mg_assert_id_like(mg_id, ..., add = NULL)

Arguments

drv

An object that inherits from DBIDriver or an existing DBIConnection (default: RPostgres::Postgres())

host

Character string giving the ip of the host to connect to

port

Host port to connect to (numeric)

dbname

Name of the database located at the host

user

Username to login with

password

Password to login with

timezone

Sets the timezone of DBI::dbConnect()

timezone_out

Sets the timezone_out of DBI::dbConnect()

...

Parameters passed to checkmate::check_*

conn

An object that inherits from DBIConnection (as generated by get_connection())'

pattern

A regex pattern with which to subset the returned tables

db_table_id

A DBI::Id() object or a character string readable by mg_id.

slice_ts

The time / date to slice by

include_slice_info

Default FALSE. If set TRUE, the history columns "checksum", "from_ts", "until_ts" are returned also

.data

Object to test if is data.table, data.frame, tbl or tibble

from_ts

The name of the column in .data specifying valid from time (note: must be unquoted)

until_ts

The name of the column in .data specifying valid until time (note: must be unquoted)

x, y

A pair of lazy data frames backed by database queries.

by

passed to mg_inner_join if different from NULL

na_by

passed to mg_inner_join if different from NULL

sql_on

A custom join predicate as an SQL expression. Usually joins use column equality, but you can perform more complex queries by supply sql_on which should be a SQL expression that uses LHS and RHS aliases to refer to the left-hand side or right-hand side of the join respectively.

renamer

named list generated by mg_select_na_sql

left

Boolean that control if joins is left (alternatively right) join

tables

A list(!) of tables you want to combine is supplied here as lazy_queries.

colnames

If the time axes of validity is not called "valid_to" and "valid_until" inside each lazy_query, you can specify their names by supplying the arguments as a list (e.g. c(t1.from = "\<colname\>", t2.until = "\<colname\>"). colnames must be named in same order as as given in tables (i.e. t1, t2, t3, ...).

col

The name of column the checksums will be placed in

exclude

Columns to exclude from the checksum generation

warn

Flag to warn if target column already exists in data

data

A data frame.

sep

Separator to use between values.

remove

If TRUE, remove input columns from output data frame.

na.rm

If TRUE, missing values will be removed prior to uniting each value.

temporary

Should the table be created as a temporary table?

db_table

Object to test if is of class "tbl_dbi" or character on form "schema.table"

timestamp

Object to test if is POSIX or character

filters

A filter to subset data by. if filters == NULL, no filtering occurs. Else, an mg_inner_join is performed using all columns of the filter

message

A message to add to the log-file (useful for supplying metadata to the log)

tic

A timestamp when computation began. If not supplied, it will be created at call-time. (Used to more accurately convey how long runtime of the update process has been)

log_path

The path where logs are stored. If NULL, no file logs are created.

log_table_id

A DBI::Id() object or a character string readable by mg_id, specifying the location of the log table.

enforce_chronological_order

A logical that controls whether or not to check if timestamp of update is prior to timestamps in the DB

log_table

A specification of where the logs should exist ("schema.table")

age_cuts

The lower bound of the groups (0 is implicitly included)

add

AssertCollection to add assertions to

mg_id

Object to test if is of class "Id" or character on form "schema.table"

Value

An object that inherits from DBIConnection driver specified in drv

A data.frame containing table names in the DB

A "lazy" dataframe (tbl_lazy) generated using dbplyr

An object similar to .data

     The combination of input queries with a single, interlaced
             valid_from / valid_until time axis

Invisibly returns the table as it looks on the destination (or locally if conn is NULL)

An object similar to .data

TRUE if .data contains the columns: "checksum", "from_ts", and "until_ts". FALSE otherwise

The number of records in the object

A vector of labels with zero-padded numerics so they can be sorted easily

Details

This function generates a much faster sql statement for NA join compared to dbplyr's _join with na_matches = "na".

See also

Public fields

log_path

(character(1))
A directory where log file is written (if this is not NULL). Defaults to getOption("mg.log_path").

log_filename

(character(1))
The name (basename) of the log file.

log_tbl

The DB table used for logging. Class is connection-specific, but inherits from tbl_dbi.

start_time

(POSIXct(1))
The time at which data processing was started.

Methods

Public methods

mg_Logger$new()

mg_Logger$log_info()

mg_Logger$log_warn()

mg_Logger$log_error()

mg_Logger$log_to_db()

mg_Logger$clone()

Method new()

Create a new mg_Logger object

Usage

mg_Logger$new(
  db_tablestring = NULL,
  log_table_id = getOption("mg.log_table_id"),
  log_conn = NULL,
  log_path = getOption("mg.log_path"),
  ts = NULL,
  start_time = Sys.time()
)

Arguments

db_tablestring

A string specifying the table being updated

log_table_id

A DBI::Id() object or a character string readable by mg_id, specifying the location of the log table.

log_conn

A database connection inheriting from DBIConnection

log_path

The path where logs are stored. If NULL, no file logs are created.

ts

A timestamp describing the data being processed (≠ current time)

start_time

The time at which data processing was started (defaults to Sys.time())

Method log_info()

Write a line to log file

Usage

mg_Logger$log_info(..., tic = Sys.time(), log_type = "INFO")

Arguments

...

One or more character strings to be concatenated

tic

The timestamp used by the log entry (default Sys.time())

log_type

A character string which describes the severity of the log message

Method log_warn()

Write a warning to log file and generate warning.

Usage

mg_Logger$log_warn(..., log_type = "WARNING")

Arguments

...

One or more character strings to be concatenated

log_type

A character string which describes the severity of the log message

Method log_error()

Write an error to log file and stop execution

Usage

mg_Logger$log_error(..., log_type = "ERROR")

Arguments

...

One or more character strings to be concatenated

log_type

A character string which describes the severity of the log message

Method log_to_db()

Write or update log table

Usage

mg_Logger$log_to_db(...)

Arguments

...

Name-value pairs with which to update the log table

Method clone()

The objects of this class are cloneable with this method.

Usage

mg_Logger$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.